Adding sensible defaults for arts
The mobile app sends us the year as an integer, but the
`escape_javascript(object)` method runs a gsub on the parameter.
For strings, this is OK, but for integers, this isn't, so we're
forcing a `to_s` for years.

The second issue is with the Flickr upload: if it fails, there isn't
an array of Flickr IDs, so the `art.flickr_ids.first` fails because
we're calling `first` on a nil object.